✨ init misc types file
