+++ modules/types/src/misc.ts
@@ -1,0 +1,44 @@
+import { Address, BigNumber, Bytes32, Xpub } from "./basic";
+import { enumify } from "./utils";
+
+export type Collateralizations = { [assetId: string]: boolean };
+
+// wtf is this?
+export interface VerifyNonceDtoType {
+  sig: string;
+  userPublicIdentifier: Xpub;
+}
+
+// used to verify channel is in sequence
+type ChannelAppSequences = {
+  userSequenceNumber: number;
+  nodeSequenceNumber: number;
+};
+
+interface PendingAsyncTransfer {
+  assetId: Address;
+  amount: BigNumber;
+  encryptedPreImage: string;
+  linkedHash: Bytes32;
+  paymentId: Bytes32;
+}
+
+////////////////////////////////////
+// Swap Rate Management
+
+export type AllowedSwap = {
+  from: Address;
+  to: Address;
+};
+
+export const PriceOracleTypes = enumify({
+  UNISWAP: "UNISWAP",
+});
+export type PriceOracleTypes = (typeof PriceOracleTypes)[keyof typeof PriceOracleTypes];
+export type PriceOracleType = keyof typeof PriceOracleTypes;
+
+export type SwapRate = AllowedSwap & {
+  rate: string; // DecString?
+  priceOracleType: PriceOracleType;
+  blockNumber?: number;
+};
